# doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Home Library Service",
    "description": "Home music library service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "login": {
            "type": "string",
            "example": "TestUser"
          },
          "version": {
            "type": "number",
            "example": 1
          },
          "createAt": {
            "type": "number",
            "example": 1655000000
          },
          "updateAt": {
            "type": "number",
            "example": 1655000000
          }
        },
        "required": [
          "id",
          "login"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get single user by id",
        "description": "Get single user by id",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. userId is invalid (not uuid)"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user's password",
        "description": "Updates a user's password by ID",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "example",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "description": "The user's old password"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "The user's new password"
                  }
                },
                "required": [
                  "oldPassword",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "example",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "login": {
                      "type": "string",
                      "example": "TestUser"
                    },
                    "version": {
                      "type": "integer",
                      "example": 2
                    },
                    "createAt": {
                      "type": "integer",
                      "example": 1655000000
                    },
                    "updateAt": {
                      "type": "integer",
                      "example": 1655999999
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. userId is invalid (not uuid)"
          },
          "403": {
            "description": "oldPassowrd is wrong"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "description": "Deletes user by ID.",
        "responses": {
          "204": {
            "description": "The user has been deleted"
          },
          "400": {
            "description": "Bad request. userId is invalid (not uuid)"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  }
}